[Style] Convert the 'position-area' property to strong style types https://bugs.webkit.org/show_bug.cgi?id=302727
Reviewed by Darin Adler.

Converts the 'position-area' property to use strong style types.

Also splits up the types previously in PositionArea.h/cpp into files
of their own.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/rendering/PositionedLayoutConstraints.cpp:
* Source/WebCore/rendering/RenderInline.cpp:
* Source/WebCore/rendering/style/PositionArea.cpp: Removed.
* Source/WebCore/rendering/style/PositionArea.h: Removed.
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleInlines.h:
* Source/WebCore/rendering/style/RenderStyleSetters.h:
* Source/WebCore/rendering/style/StyleRareNonInheritedData.h:
* Source/WebCore/style/AnchorPositionEvaluator.cpp:
* Source/WebCore/style/StyleBuilderConverter.h:
* Source/WebCore/style/StyleBuilderState.h:
* Source/WebCore/style/StyleExtractorConverter.h:
* Source/WebCore/style/StyleExtractorSerializer.h:
* Source/WebCore/style/values/anchor-position/StylePositionArea.cpp: Added.
* Source/WebCore/style/values/anchor-position/StylePositionArea.h: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaAxis.cpp: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaAxis.h: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaSelf.cpp: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaSelf.h: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaSpan.cpp: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaSpan.h: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaTrack.cpp: Added.
* Source/WebCore/style/values/anchor-position/StylePositionAreaTrack.h: Added.

Canonical link: https://commits.webkit.org/303452@main
