+++ Source/WebCore/Sources.txt
@@ -2979,7 +2979,6 @@ rendering/style/BorderValue.cpp
 rendering/style/DropShadowFilterOperationWithStyleColor.cpp
 rendering/style/GridSpan.cpp
 rendering/style/OutlineValue.cpp
-rendering/style/PositionArea.cpp
 rendering/style/PositionTryFallback.cpp
 rendering/style/PositionTryOrder.cpp
 rendering/style/ReferenceFilterOperation.cpp
@@ -3177,6 +3176,11 @@ style/values/align/StyleAlignSelf.cpp
 style/values/align/StyleJustifyContent.cpp
 style/values/align/StyleJustifyItems.cpp
 style/values/align/StyleJustifySelf.cpp
+style/values/anchor-position/StylePositionAreaAxis.cpp
+style/values/anchor-position/StylePositionAreaSelf.cpp
+style/values/anchor-position/StylePositionAreaSpan.cpp
+style/values/anchor-position/StylePositionAreaTrack.cpp
+style/values/anchor-position/StylePositionArea.cpp
 style/values/animations/StyleAnimation.cpp
 style/values/animations/StyleSingleAnimationName.cpp
 style/values/animations/StyleSingleAnimationRangeName.cpp
