+++ Source/WebCore/rendering/updating/RenderTreeUpdater.cpp
@@ -58,6 +58,7 @@
 #include "RenderTreeUpdaterViewTransition.h"
 #include "RenderView.h"
 #include "SVGElement.h"
+#include "Settings.h"
 #include "StylableInlines.h"
 #include "StyleResolver.h"
 #include "StyleTreeResolver.h"
