+++ Source/WebCore/style/StyleBuilderConverter.h
@@ -118,7 +118,6 @@ public:
     static Resize convertResize(BuilderState&, const CSSValue&);
 
     static std::optional<ScopedName> convertPositionAnchor(BuilderState&, const CSSValue&);
-    static std::optional<PositionArea> convertPositionArea(BuilderState&, const CSSValue&);
 
     static NameScope convertNameScope(BuilderState&, const CSSValue&);
 
@@ -215,365 +214,6 @@ inline std::optional<ScopedName> BuilderConverter::convertPositionAnchor(Builder
     if (value.valueID() == CSSValueAuto)
         return { };
     return convertStyleType<ScopedName>(builderState, value);
-}
-
-static std::optional<PositionAreaAxis> positionAreaKeywordToAxis(CSSValueID keyword)
-{
-    switch (keyword) {
-    case CSSValueLeft:
-    case CSSValueSpanLeft:
-    case CSSValueRight:
-    case CSSValueSpanRight:
-        return PositionAreaAxis::Horizontal;
-
-    case CSSValueTop:
-    case CSSValueSpanTop:
-    case CSSValueBottom:
-    case CSSValueSpanBottom:
-        return PositionAreaAxis::Vertical;
-
-    case CSSValueXStart:
-    case CSSValueSpanXStart:
-    case CSSValueSelfXStart:
-    case CSSValueSpanSelfXStart:
-    case CSSValueXEnd:
-    case CSSValueSpanXEnd:
-    case CSSValueSelfXEnd:
-    case CSSValueSpanSelfXEnd:
-        return PositionAreaAxis::X;
-
-    case CSSValueYStart:
-    case CSSValueSpanYStart:
-    case CSSValueSelfYStart:
-    case CSSValueSpanSelfYStart:
-    case CSSValueYEnd:
-    case CSSValueSpanYEnd:
-    case CSSValueSelfYEnd:
-    case CSSValueSpanSelfYEnd:
-        return PositionAreaAxis::Y;
-
-    case CSSValueBlockStart:
-    case CSSValueSpanBlockStart:
-    case CSSValueSelfBlockStart:
-    case CSSValueSpanSelfBlockStart:
-    case CSSValueBlockEnd:
-    case CSSValueSpanBlockEnd:
-    case CSSValueSelfBlockEnd:
-    case CSSValueSpanSelfBlockEnd:
-        return PositionAreaAxis::Block;
-
-    case CSSValueInlineStart:
-    case CSSValueSpanInlineStart:
-    case CSSValueSelfInlineStart:
-    case CSSValueSpanSelfInlineStart:
-    case CSSValueInlineEnd:
-    case CSSValueSpanInlineEnd:
-    case CSSValueSelfInlineEnd:
-    case CSSValueSpanSelfInlineEnd:
-        return PositionAreaAxis::Inline;
-
-    case CSSValueStart:
-    case CSSValueSpanStart:
-    case CSSValueSelfStart:
-    case CSSValueSpanSelfStart:
-    case CSSValueEnd:
-    case CSSValueSpanEnd:
-    case CSSValueSelfEnd:
-    case CSSValueSpanSelfEnd:
-    case CSSValueCenter:
-    case CSSValueSpanAll:
-        return { };
-
-    default:
-        ASSERT_NOT_REACHED();
-        return { };
-    }
-}
-
-static PositionAreaTrack positionAreaKeywordToTrack(CSSValueID keyword)
-{
-    switch (keyword) {
-    case CSSValueLeft:
-    case CSSValueTop:
-    case CSSValueXStart:
-    case CSSValueSelfXStart:
-    case CSSValueYStart:
-    case CSSValueSelfYStart:
-    case CSSValueBlockStart:
-    case CSSValueSelfBlockStart:
-    case CSSValueInlineStart:
-    case CSSValueSelfInlineStart:
-    case CSSValueStart:
-    case CSSValueSelfStart:
-        return PositionAreaTrack::Start;
-
-    case CSSValueSpanLeft:
-    case CSSValueSpanTop:
-    case CSSValueSpanXStart:
-    case CSSValueSpanSelfXStart:
-    case CSSValueSpanYStart:
-    case CSSValueSpanSelfYStart:
-    case CSSValueSpanBlockStart:
-    case CSSValueSpanSelfBlockStart:
-    case CSSValueSpanInlineStart:
-    case CSSValueSpanSelfInlineStart:
-    case CSSValueSpanStart:
-    case CSSValueSpanSelfStart:
-        return PositionAreaTrack::SpanStart;
-
-    case CSSValueRight:
-    case CSSValueBottom:
-    case CSSValueXEnd:
-    case CSSValueSelfXEnd:
-    case CSSValueYEnd:
-    case CSSValueSelfYEnd:
-    case CSSValueBlockEnd:
-    case CSSValueSelfBlockEnd:
-    case CSSValueInlineEnd:
-    case CSSValueSelfInlineEnd:
-    case CSSValueEnd:
-    case CSSValueSelfEnd:
-        return PositionAreaTrack::End;
-
-    case CSSValueSpanRight:
-    case CSSValueSpanBottom:
-    case CSSValueSpanXEnd:
-    case CSSValueSpanSelfXEnd:
-    case CSSValueSpanYEnd:
-    case CSSValueSpanSelfYEnd:
-    case CSSValueSpanBlockEnd:
-    case CSSValueSpanSelfBlockEnd:
-    case CSSValueSpanInlineEnd:
-    case CSSValueSpanSelfInlineEnd:
-    case CSSValueSpanEnd:
-    case CSSValueSpanSelfEnd:
-        return PositionAreaTrack::SpanEnd;
-
-    case CSSValueCenter:
-        return PositionAreaTrack::Center;
-    case CSSValueSpanAll:
-        return PositionAreaTrack::SpanAll;
-
-    default:
-        ASSERT_NOT_REACHED();
-        return PositionAreaTrack::Start;
-    }
-}
-
-static PositionAreaSelf positionAreaKeywordToSelf(CSSValueID keyword)
-{
-    switch (keyword) {
-    case CSSValueLeft:
-    case CSSValueSpanLeft:
-    case CSSValueRight:
-    case CSSValueSpanRight:
-    case CSSValueTop:
-    case CSSValueSpanTop:
-    case CSSValueBottom:
-    case CSSValueSpanBottom:
-    case CSSValueXStart:
-    case CSSValueSpanXStart:
-    case CSSValueXEnd:
-    case CSSValueSpanXEnd:
-    case CSSValueYStart:
-    case CSSValueSpanYStart:
-    case CSSValueYEnd:
-    case CSSValueSpanYEnd:
-    case CSSValueBlockStart:
-    case CSSValueSpanBlockStart:
-    case CSSValueBlockEnd:
-    case CSSValueSpanBlockEnd:
-    case CSSValueInlineStart:
-    case CSSValueSpanInlineStart:
-    case CSSValueInlineEnd:
-    case CSSValueSpanInlineEnd:
-    case CSSValueStart:
-    case CSSValueSpanStart:
-    case CSSValueEnd:
-    case CSSValueSpanEnd:
-    case CSSValueCenter:
-    case CSSValueSpanAll:
-        return PositionAreaSelf::No;
-
-    case CSSValueSelfXStart:
-    case CSSValueSpanSelfXStart:
-    case CSSValueSelfXEnd:
-    case CSSValueSpanSelfXEnd:
-    case CSSValueSelfYStart:
-    case CSSValueSpanSelfYStart:
-    case CSSValueSelfYEnd:
-    case CSSValueSpanSelfYEnd:
-    case CSSValueSelfBlockStart:
-    case CSSValueSpanSelfBlockStart:
-    case CSSValueSelfBlockEnd:
-    case CSSValueSpanSelfBlockEnd:
-    case CSSValueSelfInlineStart:
-    case CSSValueSpanSelfInlineStart:
-    case CSSValueSelfInlineEnd:
-    case CSSValueSpanSelfInlineEnd:
-    case CSSValueSelfStart:
-    case CSSValueSpanSelfStart:
-    case CSSValueSelfEnd:
-    case CSSValueSpanSelfEnd:
-        return PositionAreaSelf::Yes;
-
-    default:
-        ASSERT_NOT_REACHED();
-        return PositionAreaSelf::No;
-    }
-}
-
-// Expand a one keyword position-area to the equivalent keyword pair value.
-static std::pair<CSSValueID, CSSValueID> positionAreaExpandKeyword(CSSValueID dim)
-{
-    auto maybeAxis = positionAreaKeywordToAxis(dim);
-    if (maybeAxis) {
-        // Keyword is axis unambiguous, second keyword is span-all.
-
-        // Y/inline axis keyword goes after in the pair.
-        auto axis = *maybeAxis;
-        if (axis == PositionAreaAxis::Vertical || axis == PositionAreaAxis::Y || axis == PositionAreaAxis::Inline)
-            return { CSSValueSpanAll, dim };
-
-        return { dim, CSSValueSpanAll };
-    }
-
-    // Keyword is axis ambiguous, it's repeated.
-    return { dim, dim };
-}
-
-
-// Flip a PositionArea across a logical axis (block or inline), given the current writing mode.
-inline PositionArea flipPositionAreaByLogicalAxis(LogicalBoxAxis flipAxis, PositionArea area, WritingMode writingMode)
-{
-    auto blockOrXSpan = area.blockOrXAxis();
-    auto inlineOrYSpan = area.inlineOrYAxis();
-
-    // blockOrXSpan is on the flip axis, so flip its track and keep inlineOrYSpan intact.
-    if (mapPositionAreaAxisToLogicalAxis(blockOrXSpan.axis(), writingMode) == flipAxis) {
-        return {
-            { blockOrXSpan.axis(), flipPositionAreaTrack(blockOrXSpan.track()), blockOrXSpan.self() },
-            inlineOrYSpan
-        };
-    }
-
-    // The two spans are orthogonal in axis, so if blockOrXSpan isn't on the flip axis,
-    // inlineOrYSpan must be. In this case, flip the track of inlineOrYSpan, and
-    // keep blockOrXSpan intact.
-    return {
-        blockOrXSpan,
-        { inlineOrYSpan.axis(), flipPositionAreaTrack(inlineOrYSpan.track()), inlineOrYSpan.self() }
-    };
-}
-
-// Flip a PositionArea across a physical axis (x or y), given the current writing mode.
-inline PositionArea flipPositionAreaByPhysicalAxis(BoxAxis flipAxis, PositionArea area, WritingMode writingMode)
-{
-    auto blockOrXSpan = area.blockOrXAxis();
-    auto inlineOrYSpan = area.inlineOrYAxis();
-
-    // blockOrXSpan is on the flip axis, so flip its track and keep inlineOrYSpan intact.
-    if (mapPositionAreaAxisToPhysicalAxis(blockOrXSpan.axis(), writingMode) == flipAxis) {
-        return {
-            { blockOrXSpan.axis(), flipPositionAreaTrack(blockOrXSpan.track()), blockOrXSpan.self() },
-            inlineOrYSpan
-        };
-    }
-
-    // The two spans are orthogonal in axis, so if blockOrXSpan isn't on the flip axis,
-    // inlineOrYSpan must be. In this case, flip the track of inlineOrYSpan, and
-    // keep blockOrXSpan intact.
-    return {
-        blockOrXSpan,
-        { inlineOrYSpan.axis(), flipPositionAreaTrack(inlineOrYSpan.track()), inlineOrYSpan.self() }
-    };
-}
-
-// Flip a PositionArea as specified by flip-start tactic.
-// Intuitively, this mirrors the PositionArea across a diagonal line drawn from the
-// block-start/inline-start corner to the block-end/inline-end corner. This is done
-// by flipping the axes of the spans in the PositionArea, while keeping their track
-// and self properties intact. Because this turns a block/X span into an inline/Y
-// span and vice versa, this function also swaps the order of the spans, so
-// that the block/X span goes before the inline/Y span.
-inline PositionArea mirrorPositionAreaAcrossDiagonal(PositionArea area)
-{
-    auto blockOrXSpan = area.blockOrXAxis();
-    auto inlineOrYSpan = area.inlineOrYAxis();
-
-    return {
-        { oppositePositionAreaAxis(inlineOrYSpan.axis()), inlineOrYSpan.track(), inlineOrYSpan.self() },
-        { oppositePositionAreaAxis(blockOrXSpan.axis()), blockOrXSpan.track(), blockOrXSpan.self() }
-    };
-}
-
-inline std::optional<PositionArea> BuilderConverter::convertPositionArea(BuilderState& builderState, const CSSValue& value)
-{
-    std::pair<CSSValueID, CSSValueID> dimPair;
-
-    if (value.isValueID()) {
-        if (value.valueID() == CSSValueNone)
-            return { };
-
-        dimPair = positionAreaExpandKeyword(value.valueID());
-    } else if (const auto* pair = dynamicDowncast<CSSValuePair>(value)) {
-        const auto& first = pair->first();
-        const auto& second = pair->second();
-        ASSERT(first.isValueID() && second.isValueID());
-
-        // The parsing logic guarantees the keyword pair is in the correct order
-        // (horizontal/x/block axis before vertical/Y/inline axis)
-
-        dimPair = { first.valueID(), second.valueID() };
-    } else {
-        // value MUST be a single ValueID or a pair of ValueIDs, as returned by the parsing logic.
-        ASSERT_NOT_REACHED();
-        return { };
-    }
-
-    auto dim1Axis = positionAreaKeywordToAxis(dimPair.first);
-    auto dim2Axis = positionAreaKeywordToAxis(dimPair.second);
-
-    // If both keyword axes are ambiguous, the first one is block axis and second one
-    // is inline axis. If only one keyword axis is ambiguous, its axis is the opposite
-    // of the other keyword's axis.
-    if (!dim1Axis && !dim2Axis) {
-        dim1Axis = PositionAreaAxis::Block;
-        dim2Axis = PositionAreaAxis::Inline;
-    } else if (!dim1Axis)
-        dim1Axis = oppositePositionAreaAxis(*dim2Axis);
-    else if (!dim2Axis)
-        dim2Axis = oppositePositionAreaAxis(*dim1Axis);
-
-    PositionArea area {
-        { *dim1Axis, positionAreaKeywordToTrack(dimPair.first), positionAreaKeywordToSelf(dimPair.first) },
-        { *dim2Axis, positionAreaKeywordToTrack(dimPair.second), positionAreaKeywordToSelf(dimPair.second) }
-    };
-
-    // Flip according to position-try-fallbacks, if specified.
-    if (const auto& positionTryFallback = builderState.positionTryFallback()) {
-        for (auto tactic : positionTryFallback->tactics) {
-            switch (tactic) {
-            case PositionTryFallback::Tactic::FlipBlock:
-                area = flipPositionAreaByLogicalAxis(LogicalBoxAxis::Block, area, builderState.style().writingMode());
-                break;
-            case PositionTryFallback::Tactic::FlipInline:
-                area = flipPositionAreaByLogicalAxis(LogicalBoxAxis::Inline, area, builderState.style().writingMode());
-                break;
-            case PositionTryFallback::Tactic::FlipX:
-                area = flipPositionAreaByPhysicalAxis(BoxAxis::Horizontal, area, builderState.style().writingMode());
-                break;
-            case PositionTryFallback::Tactic::FlipY:
-                area = flipPositionAreaByPhysicalAxis(BoxAxis::Vertical, area, builderState.style().writingMode());
-                break;
-            case PositionTryFallback::Tactic::FlipStart:
-                area = mirrorPositionAreaAcrossDiagonal(area);
-                break;
-            }
-        }
-    }
-
-    return area;
 }
 
 inline NameScope BuilderConverter::convertNameScope(BuilderState& builderState, const CSSValue& value)
