+++ Source/WebCore/style/StyleExtractorConverter.h
@@ -133,8 +133,6 @@ public:
 
     static Ref<CSSValue> convertPositionTryFallbacks(ExtractorState&, const FixedVector<PositionTryFallback>&);
     static Ref<CSSValue> convertPositionAnchor(ExtractorState&, const std::optional<ScopedName>&);
-    static Ref<CSSValue> convertPositionArea(ExtractorState&, const PositionArea&);
-    static Ref<CSSValue> convertPositionArea(ExtractorState&, const std::optional<PositionArea>&);
     static Ref<CSSValue> convertNameScope(ExtractorState&, const NameScope&);
 
     // MARK: MaskLayer property conversions
@@ -264,148 +262,6 @@ inline Ref<CSSValue> ExtractorConverter::convertPositionAnchor(ExtractorState& s
     if (!positionAnchor)
         return CSSPrimitiveValue::create(CSSValueAuto);
     return convert(state, *positionAnchor);
-}
-
-inline Ref<CSSValue> ExtractorConverter::convertPositionArea(ExtractorState&, const PositionArea& positionArea)
-{
-    auto keywordForPositionAreaSpan = [](const PositionAreaSpan span) -> CSSValueID {
-        auto axis = span.axis();
-        auto track = span.track();
-        auto self = span.self();
-
-        switch (axis) {
-        case PositionAreaAxis::Horizontal:
-            ASSERT(self == PositionAreaSelf::No);
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return CSSValueLeft;
-            case PositionAreaTrack::SpanStart:
-                return CSSValueSpanLeft;
-            case PositionAreaTrack::End:
-                return CSSValueRight;
-            case PositionAreaTrack::SpanEnd:
-                return CSSValueSpanRight;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueLeft;
-            }
-
-        case PositionAreaAxis::Vertical:
-            ASSERT(self == PositionAreaSelf::No);
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return CSSValueTop;
-            case PositionAreaTrack::SpanStart:
-                return CSSValueSpanTop;
-            case PositionAreaTrack::End:
-                return CSSValueBottom;
-            case PositionAreaTrack::SpanEnd:
-                return CSSValueSpanBottom;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueTop;
-            }
-
-        case PositionAreaAxis::X:
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return self == PositionAreaSelf::No ? CSSValueXStart : CSSValueSelfXStart;
-            case PositionAreaTrack::SpanStart:
-                return self == PositionAreaSelf::No ? CSSValueSpanXStart : CSSValueSpanSelfXStart;
-            case PositionAreaTrack::End:
-                return self == PositionAreaSelf::No ? CSSValueXEnd : CSSValueSelfXEnd;
-            case PositionAreaTrack::SpanEnd:
-                return self == PositionAreaSelf::No ? CSSValueSpanXEnd : CSSValueSpanSelfXEnd;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueXStart;
-            }
-
-        case PositionAreaAxis::Y:
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return self == PositionAreaSelf::No ? CSSValueYStart : CSSValueSelfYStart;
-            case PositionAreaTrack::SpanStart:
-                return self == PositionAreaSelf::No ? CSSValueSpanYStart : CSSValueSpanSelfYStart;
-            case PositionAreaTrack::End:
-                return self == PositionAreaSelf::No ? CSSValueYEnd : CSSValueSelfYEnd;
-            case PositionAreaTrack::SpanEnd:
-                return self == PositionAreaSelf::No ? CSSValueSpanYEnd : CSSValueSpanSelfYEnd;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueYStart;
-            }
-
-        case PositionAreaAxis::Block:
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return self == PositionAreaSelf::No ? CSSValueBlockStart : CSSValueSelfBlockStart;
-            case PositionAreaTrack::SpanStart:
-                return self == PositionAreaSelf::No ? CSSValueSpanBlockStart : CSSValueSpanSelfBlockStart;
-            case PositionAreaTrack::End:
-                return self == PositionAreaSelf::No ? CSSValueBlockEnd : CSSValueSelfBlockEnd;
-            case PositionAreaTrack::SpanEnd:
-                return self == PositionAreaSelf::No ? CSSValueSpanBlockEnd : CSSValueSpanSelfBlockEnd;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueBlockStart;
-            }
-
-        case PositionAreaAxis::Inline:
-            switch (track) {
-            case PositionAreaTrack::Start:
-                return self == PositionAreaSelf::No ? CSSValueInlineStart : CSSValueSelfInlineStart;
-            case PositionAreaTrack::SpanStart:
-                return self == PositionAreaSelf::No ? CSSValueSpanInlineStart : CSSValueSpanSelfInlineStart;
-            case PositionAreaTrack::End:
-                return self == PositionAreaSelf::No ? CSSValueInlineEnd : CSSValueSelfInlineEnd;
-            case PositionAreaTrack::SpanEnd:
-                return self == PositionAreaSelf::No ? CSSValueSpanInlineEnd : CSSValueSpanSelfInlineEnd;
-            case PositionAreaTrack::Center:
-                return CSSValueCenter;
-            case PositionAreaTrack::SpanAll:
-                return CSSValueSpanAll;
-            default:
-                ASSERT_NOT_REACHED();
-                return CSSValueInlineStart;
-            }
-        }
-
-        ASSERT_NOT_REACHED();
-        return CSSValueLeft;
-    };
-
-    auto blockOrXAxisKeyword = keywordForPositionAreaSpan(positionArea.blockOrXAxis());
-    auto inlineOrYAxisKeyword = keywordForPositionAreaSpan(positionArea.inlineOrYAxis());
-
-    return CSSPropertyParserHelpers::valueForPositionArea(blockOrXAxisKeyword, inlineOrYAxisKeyword, CSSPropertyParserHelpers::ValueType::Computed).releaseNonNull();
-}
-
-inline Ref<CSSValue> ExtractorConverter::convertPositionArea(ExtractorState& state, const std::optional<PositionArea>& positionArea)
-{
-    if (!positionArea)
-        return CSSPrimitiveValue::create(CSSValueNone);
-    return convertPositionArea(state, *positionArea);
 }
 
 inline Ref<CSSValue> ExtractorConverter::convertNameScope(ExtractorState&, const NameScope& scope)
