+++ Source/WebCore/style/values/anchor-position/StylePositionAreaTrack.cpp
@@ -1,0 +1,48 @@
+/*
+ * Copyright (C) 2025 Samuel Weinig <sam@webkit.org>
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+ * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+#include "StylePositionAreaTrack.h"
+
+#include <wtf/text/TextStream.h>
+
+namespace WebCore {
+namespace Style {
+
+WTF::TextStream& operator<<(WTF::TextStream& ts, PositionAreaTrack track)
+{
+    switch (track) {
+    case PositionAreaTrack::Start:     ts << "start"_s; break;
+    case PositionAreaTrack::SpanStart: ts << "span-start"_s; break;
+    case PositionAreaTrack::End:       ts << "end"_s; break;
+    case PositionAreaTrack::SpanEnd:   ts << "span-end"_s; break;
+    case PositionAreaTrack::Center:    ts << "center"_s; break;
+    case PositionAreaTrack::SpanAll:   ts << "span-all"_s; break;
+    }
+
+    return ts;
+}
+
+} // namespace Style
+} // namespace WebCore
